+++ bigtree/tree/tree.py
@@ -24,7 +24,8 @@ class Tree:
 
         * Construction with dataframe, dictionary, list, or string
         * Export to dataframe, dictionary, list, string, or images
-        * Helper methods for cloning, pruning, getting tree diff
+        * Helper methods for cloning, pruning, getting subtree and tree diff
+        * Iterator methods to parse tree
         * Query and Search methods to find one or more Nodes
         * Plot methods
 
@@ -66,7 +67,10 @@ class Tree:
         else:
 
             def wrapper(self, other_tree: T, *args, **kwargs):  # type: ignore
-                return func(self.node, other_tree.node, *args, **kwargs)
+                results = func(self.node, other_tree.node, *args, **kwargs)
+                if isinstance(results, node.Node):
+                    return type(self)(results)
+                return results
 
         functools.update_wrapper(wrapper, base_func)
         wrapper.__name__ = name